+++ Documentation/Changelog.md
@@ -8,7 +8,7 @@ Release date: not released yet
 
 - Polyfills:
     - Added polyfill for `System.Security.Cryptography.CryptographicOperations` class
-    - Added polyfills for `Fill(Span<byte>)`, `GetBytes(Span<byte>)`, and `GetBytes(int)` methods of `System.Security.Cryptography.RandomNumberGenerator` class
+    - Added polyfills for `Fill`, `GetBytes`, and `GetInt32` methods of `System.Security.Cryptography.RandomNumberGenerator` class
 - Fixed the signature of `Process.WaitForExit(int, CancellationToken)` extension method to return a `bool` result
 
 ### Gapotchenko.FX 2026.7